LayoutTests/imported/w3c/web-platform-tests/html/syntax/parsing/template/additions-to-foster-parenting/template-is-not-a-foster-parent-element-expected.txt


PASS Template is not a foster parent element. Test the case when <template> is higher in stack of open elements
PASS Template is not a foster parent element. Test the case when <template> is lower in stack of open elements
